+++ src/lib.rs
@@ -11,7 +11,7 @@ pub(crate) mod validation;
 
 pub use error::{LinkbaseType, Result, XbrlError};
 pub use instance::{
-    Context, ContextId, DocumentView, EntityIdentifier, Fact, FootnoteArc, FootnoteLink,
+    Context, ContextId, Decimals, DocumentView, EntityIdentifier, Fact, FootnoteArc, FootnoteLink,
     FootnoteLocator, FootnoteResource, NamespacePrefix, Period, SectionView, TreeNode, Unit,
     UnitId, XbrlInstance,
 };
@@ -19,8 +19,9 @@ pub use quick_xml::{Reader as XmlReader, Writer as XmlWriter};
 #[cfg(feature = "download")]
 pub use taxonomy::TaxonomyLoader;
 pub use taxonomy::{
-    ArcroleType, CalculationArc, ConceptId, DefinitionArc, ElementDefinition, Label,
-    LinkbaseLocator, LinkbaseRef, PresentationArc, Reference, ReferencePart, RoleType, RoleUri,
-    SchemaImport, SchemaInclude, SchemaRefUrl, TaxonomySchema, TaxonomySet,
+    ArcroleType, Balance, CalculationArc, ConceptId, CyclesAllowed, DefinitionArc,
+    ElementDefinition, Label, LinkbaseLocator, LinkbaseRef, PeriodType, PresentationArc, Reference,
+    ReferencePart, RoleType, RoleUri, SchemaImport, SchemaInclude, SchemaRefUrl, TaxonomySchema,
+    TaxonomySet,
 };
 pub use validation::{Severity, ValidationMessage, ValidationResult};